refactor: extract AccessibilityNodeInfo extensions from focus controller
